+++ Models/Config/OrganizationModel.cs
@@ -1,0 +1,7 @@
+namespace etvctl.Models.Config;
+
+public class OrganizationModel
+{
+    public FileOrganization? Default { get; set; }
+    public OrganizationOverridesModel? Overrides { get; set; }
+}