+++ Models/Config/OrganizationOverridesModel.cs
@@ -1,0 +1,9 @@
+using YamlDotNet.Serialization;
+
+namespace etvctl.Models.Config;
+
+public class OrganizationOverridesModel
+{
+    [YamlMember(Alias = "smart_collection", ApplyNamingConventions = false)]
+    public FileOrganization? SmartCollection { get; set; }
+}